[Frontend] Relationship distribution list widget refetch (#14448)
